final build

+++ app/Http/Controllers/ScheduleController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ScheduleSetRequest;
 use App\Schedule;
 use Binance;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 
 class ScheduleController extends Controller
@@ -19,6 +20,15 @@ class ScheduleController extends Controller
 
 
     public function set(Request $request){
+
+        $request->validate([
+            'symbol' => 'required',
+            'side' => 'required',
+            'amount' => 'required',
+            'sequence' => 'required',
+            'minutes' => 'required',
+            'time' => 'required',
+        ]);
 
         Schedule::create([
             'user_id' => Auth::user()->id,
@@ -35,14 +45,24 @@ class ScheduleController extends Controller
     }
 
     public function delete(Request $request){
-        $sched = Schedule::find($request->id);
-        $sched->delete();
 
+        $sched = Schedule::find($request->id);
+
+        if(auth()->user()->id !== $sched->user_id) {
+            return response()->json(['error' => 'Not authorized.'],403);
+        }
+
+        $sched->delete();
         return redirect('schedule');
     }
 
     public function activate(Request $request){
         $sched = Schedule::find($request->id);
+
+        if(auth()->user()->id !== $sched->user_id) {
+            return response()->json(['error' => 'Not authorized.'],403);
+        }
+
         $sched->status = 1;
         $sched->update();
 
@@ -51,6 +71,11 @@ class ScheduleController extends Controller
 
     public function deactivate(Request $request){
         $sched = Schedule::find($request->id);
+
+        if(auth()->user()->id !== $sched->user_id) {
+            return response()->json(['error' => 'Not authorized.'],403);
+        }
+
         $sched->status = 0;
         $sched->update();
